backend/requirements.txt
tqdm
fastapi
uvicorn
qdrant-client[fastembed]>=1.8.2
openai-whisper
openwakeword